+++ organic-search-analytics/inc/html/_head.php
@@ -10,11 +10,32 @@ if( !strpos( $_SERVER['SCRIPT_NAME'], "upgrade.php" ) ) {
 		<title><?php echo $titleTag; ?></title>
 
 		<link rel="stylesheet" href="css/styles.css">
-		<link rel="stylesheet" href="//code.jquery.com/ui/1.11.2/themes/smoothness/jquery-ui.css">
+		<link rel="stylesheet" href="css/lib/jquery/jquery-ui-1.11.2.css">
 
-		<script src="//code.jquery.com/jquery-1.10.2.js"></script>
-		<script src="//code.jquery.com/ui/1.11.2/jquery-ui.js"></script>
+		<!--[if lt IE 9]><script language="javascript" type="text/javascript" src="js/lib/jqplot/excanvas.js"></script><![endif]-->
+		<script language="javascript" type="text/javascript" src="js/lib/jquery/jquery-1.11.3.min.js"></script>
+		<script src="js/lib/jquery/jquery-ui-1.11.2.js"></script>
 		<script type="text/javascript" src="js/script.js"></script>
+
+		<?php if( isset( $GLOBALS['scriptName'] ) && $GLOBALS['scriptName'] == "report-custom.php" ) { ?>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/jquery.jqplot.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.cursor.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.dateAxisRenderer.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.canvasAxisLabelRenderer.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.canvasAxisTickRenderer.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.canvasTextRenderer.min.js"></script>
+
+
+<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.highlighter.min.js"></script>
+
+
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.barRenderer.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.categoryAxisRenderer.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.dragable.min.js"></script>
+			<script language="javascript" type="text/javascript" src="js/lib/jqplot/plugins/jqplot.trendline.min.js"></script>
+		<?php } ?>
+		<link rel="stylesheet" type="text/css" href="css/lib/jqplot/jquery.jqplot.css" />
+
 
 		<?php if( isset( $GLOBALS['scriptName'] ) && $GLOBALS['scriptName'] == "settings.php" ) { ?>
 			<script type="text/javascript" src="js/settings.js"></script>