+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## 2.0.0 - 2015-09-14
+## Unreleased
+### Changed
+- Reports - first domain is checked by default
+- Reports - added query match type (exact or broad)
+- jQuery and jQuery UI added to the code repository as opposed to loading from 3rd party servers
+
+### Added
+- jqPlot library
+- Reports - Graph added
+- Reports - Sort by
+- Reports - Sort Direction
+
+## 2.0.2 - 2015-09-14
 ### Added
 - Bing Webmaster Tools integration.
 - Import Bing Search Keywords to the database.  Bing groups their keyword data on a weekly basis.